app/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
gTTS
requests
pydantic
